[TNT-179] refactor: 도메인 모델명 수정 및 패키지 이동
