+++ data/repository/src/main/java/co/kr/data/repository/TrainerRepositoryImpl.kt
@@ -2,8 +2,8 @@ package co.kr.data.repository
 
 import co.kr.data.network.model.trainer.toDomain
 import co.kr.data.network.source.TrainerRemoteDataSource
-import co.kr.tnt.domain.model.trainer.DailyPtSession
-import co.kr.tnt.domain.model.trainer.DailyPtSessionCount
+import co.kr.tnt.domain.model.trainer.TrainerDailyPtSession
+import co.kr.tnt.domain.model.trainer.TrainerDailyPtSessionCount
 import co.kr.tnt.domain.repository.TrainerRepository
 import co.kr.tnt.domain.utils.DateFormatter
 import java.time.LocalDate
@@ -16,7 +16,7 @@ internal class TrainerRepositoryImpl @Inject constructor(
     private val trainerRemoteDataSource: TrainerRemoteDataSource,
     private val dateFormatter: DateFormatter,
 ) : TrainerRepository {
-    override suspend fun getMonthlyPtSessionCounts(yearMonth: YearMonth): List<DailyPtSessionCount> =
+    override suspend fun getMonthlyPtSessionCounts(yearMonth: YearMonth): List<TrainerDailyPtSessionCount> =
         trainerRemoteDataSource.getMonthlyPtSessionCounts(
             year = yearMonth.year,
             month = yearMonth.monthValue,
@@ -24,7 +24,7 @@ internal class TrainerRepositoryImpl @Inject constructor(
             response.toDomain(dateFormatter)
         }
 
-    override suspend fun getDailyPtSessions(day: LocalDate): DailyPtSession =
+    override suspend fun getDailyPtSessions(day: LocalDate): TrainerDailyPtSession =
         trainerRemoteDataSource.getDailyPtSessions(
             date = dateFormatter.format(day, "yyyy-MM-dd"),
         ).toDomain(dateFormatter)
