+++ domain/src/main/java/co/kr/tnt/domain/model/trainer/TrainerDailyPtSession.kt
@@ -1,0 +1,9 @@
+package co.kr.tnt.domain.model.trainer
+
+import co.kr.tnt.domain.model.PtSession
+import java.time.LocalDate
+
+data class TrainerDailyPtSession(
+    val date: LocalDate,
+    val sessions: List<PtSession>,
+)
