+++ domain/src/main/java/co/kr/tnt/domain/model/trainer/TrainerDailyPtSessionCount.kt
@@ -2,7 +2,7 @@ package co.kr.tnt.domain.model.trainer
 
 import java.time.LocalDate
 
-data class DailyPtSessionCount(
+data class TrainerDailyPtSessionCount(
     val date: LocalDate,
     val count: Int,
 )
