+++ domain/src/main/java/co/kr/tnt/domain/repository/TrainerRepository.kt
@@ -1,12 +1,12 @@
 package co.kr.tnt.domain.repository
 
-import co.kr.tnt.domain.model.trainer.DailyPtSession
-import co.kr.tnt.domain.model.trainer.DailyPtSessionCount
+import co.kr.tnt.domain.model.trainer.TrainerDailyPtSession
+import co.kr.tnt.domain.model.trainer.TrainerDailyPtSessionCount
 import java.time.LocalDate
 import java.time.YearMonth
 
 interface TrainerRepository {
-    suspend fun getMonthlyPtSessionCounts(yearMonth: YearMonth): List<DailyPtSessionCount>
+    suspend fun getMonthlyPtSessionCounts(yearMonth: YearMonth): List<TrainerDailyPtSessionCount>
 
-    suspend fun getDailyPtSessions(day: LocalDate): DailyPtSession
+    suspend fun getDailyPtSessions(day: LocalDate): TrainerDailyPtSession
 }
